Allow excessive float precision
We could also truncate the float literals to prevent the warnings. However, these constants are defined with full precision in the original implementation. Even though the extra digits are dropped by the compiler, the full number may prove useful if the model switches to long doubles at some point or if one wants to look up physical constants by value in the source code.